Create Huizar_FinalProject.py
This is what I have completed for the due date. I was working on adaptive step sizes (and other methods, starting from scratch) and they did not help the final positions. I know it is not everything that was necessary, but if it okay with you I'd like to continue working on it before Saturday so I have something worth presenting to the class.